requirements.txt
aiogram>=3.4.0
google-genai>=0.3.0